% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtcatch_handle_indir.R
\name{make_param}
\alias{make_param}
\title{Check input directory name}
\usage{
make_param(datdir, year_ad, month, pref_name)
}
\arguments{
\item{datdir}{<- "../data/"}

\item{year_ad}{<- 2018 # Not fiscal}

\item{month}{<- 10}

\item{pref_name}{<- "kagoshima"}
}
\value{
String of input directory
}
\description{
Check input directory name
}
\examples{
check_dirname("../data/", 2018, 10, "kagoshima")

}
